back_end/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Mail==0.9.1
python-dotenv==0.20.0
waitress==2.1.2
PyJWT==2.6.0
black==22.10.0
mysql-connector-python==8.0.31
requests==2.28.1
numpy==1.23.4
cryptography==38.0.4
pycryptodome==3.16.0
flask-talisman==1.0.0
Pillow==9.4.0
gunicorn==20.1.0
pyotp==2.8.0
qrcode==7.3.1
mysqlclient==2.1.1
opencv-python==4.7.0.68